requirements.txt
django
djangorestframework
python-dotenv
django-debug-toolbar